added new buff config

+++ src/main/java/com/gmail/markushygedombrowski/levels/LevelUpListener.java
@@ -32,7 +32,7 @@ public class LevelUpListener implements Listener {
         p.sendMessage("Tillykke du er nu level §b" + profile.getProperty("level") + "§3!");
         p.sendMessage("Du skal bruge §b" + profile.getXpToNextLvl() + " §3exp til næste level");
         p.sendMessage("§6§l--------§a§lLevel Up!§6§l--------");
-        levelRewards.giveReward(p, (int)profile.getProperty("level"), profile);
+        levelRewards.giveReward(p, ((Double)profile.getProperty("level")).intValue(), profile);
         try {
             playerProfiles.save(profile);
         } catch (SQLException e) {
